% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{apply_base_sd}
\alias{apply_base_sd}
\title{Apply a base-level standard deviation to a stocks uncertainty dataframe}
\usage{
apply_base_sd(uncertainty_df, base_sd)
}
\arguments{
\item{uncertainty_df}{A dataframe containing the uncertainty for an input stock}

\item{base_sd}{A numeric base-level standard deviation to apply to the stock}
}
\value{
The input dataframe with the base-level standard deviation applied
}
\description{
Apply a base-level standard deviation to a stocks uncertainty dataframe
}
