% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{derive_flow_uncertainty}
\alias{derive_flow_uncertainty}
\title{Derive uncertainty (sd) for a flow using the flow counts and flow cv along with
a base sd}
\usage{
derive_flow_uncertainty(flow_df, flow_uncertainty_df, flow_base_sd)
}
\arguments{
\item{flow_df}{A dataframe containing the flow counts}

\item{flow_uncertainty_df}{A dataframe containing the flow cv}

\item{flow_base_sd}{A numeric representation of the base level uncertainty (sd)}
}
\value{
A dataframe with the derived sd
}
\description{
Derive uncertainty (sd) for a flow using the flow counts and flow cv along with
a base sd
}
