% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{prepare_rates}
\alias{prepare_rates}
\title{Prepare rates for DPM input, forcing uppercase 'sex' and subsetting by age}
\usage{
prepare_rates(rates, rates_age_selection = c(0:105))
}
\arguments{
\item{rates}{Dataframe containing raw imported rates data}

\item{rates_age_selection}{A vector containing the ages to subset rates to}

\item{columns}{A vector containing columns to subset to for input into TMB DPM (default: region, age, sex, time, rate)}
}
\value{
A dataframe containing the required standardised rates
}
\description{
Prepare rates for DPM input, forcing uppercase 'sex' and subsetting by age
}
