% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{setup_input_counts}
\alias{setup_input_counts}
\title{Import, filter and standardise input stocks}
\usage{
setup_input_counts(
  input_data_directory,
  filenames,
  time_selection,
  age_selection = c(0:105)
)
}
\arguments{
\item{input_data_directory}{String filepath to the data directory storing stocks .csv files}

\item{filenames}{Vector containing filenames of stocks to setup}
}
\value{
A dataframe containing the imported, filtered and standardised stocks
}
\description{
Import, filter and standardise input stocks
}
