% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{setup_input_rates}
\alias{setup_input_rates}
\title{Set up the input rates for the TMB DPM, reading the .csv file, subsetting to
expected columns, standardising 'sex' and subsetting to an age range}
\usage{
setup_input_rates(
  input_data_directory,
  rates_filename,
  rates_age_selection = c(0:105)
)

setup_input_rates(
  input_data_directory,
  rates_filename,
  rates_age_selection = c(0:105)
)
}
\arguments{
\item{input_data_directory}{String filepath to the data directory}

\item{rates_filename}{String filename of the rates file to setup}

\item{rates_age_selection}{A vector containing the ages to subset rates to}
}
\value{
A dataframe containing the required standardised rates

A dataframe containing the required standardised rates
}
\description{
Set up the input rates for the TMB DPM, reading the .csv file, subsetting to
expected columns, standardising 'sex' and subsetting to an age range

Set up the input rates for the TMB DPM, reading the .csv file, subsetting to
expected columns, standardising 'sex' and subsetting to an age range
}
