% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{setup_input_stocks_uncertainty}
\alias{setup_input_stocks_uncertainty}
\title{Import, filter and standardise input stocks uncertainty}
\usage{
setup_input_stocks_uncertainty(
  input_data_directory,
  stock_uncertainty_filenames,
  stock_time_selection,
  stocks_df,
  stocks_base_sd
)
}
\arguments{
\item{input_data_directory}{String filepath to the data directory storing stocks .csv files}

\item{stock_uncertainty_filenames}{Vector containing filenames of stocks uncertainty to setup}

\item{stock_time_selection}{A int vector containing years to subset stocks uncertainty time range}

\item{stocks_df}{A dataframe containing related stocks for uncertainty}
}
\value{
A dataframe containing the imported, filtered and standardised stocks uncertainty
}
\description{
Import, filter and standardise input stocks uncertainty
}
