% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_functions.R
\name{standardise_counts}
\alias{standardise_counts}
\title{Apply consistent standardization to DPM stocks/flows data counts and sex (uppercase)}
\usage{
standardise_counts(counts_df, time_selection, age_selection = c(0:105))
}
\arguments{
\item{counts_df}{}
}
\value{
A dataframe with standardized count and sex columns
}
\description{
Apply consistent standardization to DPM stocks/flows data counts and sex (uppercase)
}
