% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{update_runtime_log}
\alias{update_runtime_log}
\title{Append running information to a log file, including a text string 'tag', a time,
a time difference and the log filepath}
\usage{
update_runtime_log(tag, time, time_diff, log_filepath)
}
\arguments{
\item{tag}{String log tag}

\item{time}{Current time}

\item{time_diff}{Time difference to log}

\item{log_filepath}{Path to log file}
}
\value{
Input log file with log objects information appended
}
\description{
Append running information to a log file, including a text string 'tag', a time,
a time difference and the log filepath
}
