% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{write_config_log}
\alias{write_config_log}
\title{Print all variables from the dpm_config.R script into a provided log file}
\usage{
write_config_log(log_filepath)
}
\arguments{
\item{log_filepath}{Path to log file}
}
\value{
Input log file with config parameters appended
}
\description{
Print all variables from the dpm_config.R script into a provided log file
}
