% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging_functions.R
\name{write_setup_log}
\alias{write_setup_log}
\title{Print variables from the dpm_setup_tmb.R script into a provided log file}
\usage{
write_setup_log(log_filepath)
}
\arguments{
\item{log_filepath}{Path to log file}
}
\value{
Input log file with setup objects information appended
}
\description{
Print variables from the dpm_setup_tmb.R script into a provided log file
}
